Refactored Card
Add a card Details components, because I since each card will have
different detail based on data, I thought a component handle that would
be better than having the rendering and determining of what details
goes into a card, all being handled by one component,also didn't want to
create to separate functions to handle this use-case because it would
break standard principle for working react,
so essentially for separation of concerns and singleton property.
Moved what was in the cards component
to the card Details component.

To do:
1. Still need to figure out how im gonna import the data into the
component.
2. For each data inside in the data object grab the values needed and
assign them to the proper tags(html container).
And this has to also account for which view is currently active.